fix(components): address audit findings - v-model consistency, component size, lint errors
- Fix NyxSelect v-model naming: use model instead of modelValue with normalisedModel computed
- Extract keyboard navigation from NyxSelect into useSelectKeyboardControls composable
- Extract useEditorAnnotations and useEditorMeta from NyxEditor to composables/
- Use arrow function pattern for composables (const instead of function)
- Reduce NyxSelect from 342 to 272 lines (within 300-line limit)
- Fix NyxEditor lint errors: unused imports, unused vars, any types in tests
- Update AGENTS.md with v-model naming convention and 300-line component limit
- Move NyxLoader implementation to classes/ (was just re-export from utils)
