+++ package.json
@@ -2,7 +2,7 @@
   "name": "nyx-kit",
   "homepage": "http://nyxkit.github.io/nyx-kit",
   "author": "Arne Decant <hello@arnedecant.be>",
-  "version": "2.0.10",
+  "version": "2.0.11",
   "packageManager": "pnpm@10.18.3",
   "private": false,
   "license": "MIT",
