+++ src/components/index.ts
@@ -1,4 +1,3 @@
-import NyxEditor from './NyxEditor/NyxEditor.vue'
 import NyxAvatar from './NyxAvatar/NyxAvatar.vue'
 import NyxBadge from './NyxBadge/NyxBadge.vue'
 import NyxBreadcrumbs from './NyxBreadcrumbs/NyxBreadcrumbs.vue'
@@ -6,9 +5,10 @@ import NyxButton from './NyxButton/NyxButton.vue'
 import NyxCard from './NyxCard/NyxCard.vue'
 import NyxCarousel from './NyxCarousel/NyxCarousel.vue'
 import NyxCheckbox from './NyxCheckbox/NyxCheckbox.vue'
+import { NyxEditor } from './NyxEditor/index'
 import NyxForm from './NyxForm/NyxForm.vue'
 import NyxFormField from './NyxForm/NyxFormField.vue'
-import NyxGrid from './NyxGrid/NyxGrid.vue'
+import { NyxGrid } from './NyxGrid/index'
 import NyxInput from './NyxInput/NyxInput.vue'
 import NyxMedia from './NyxMedia/NyxMedia.vue'
 import NyxModal from './NyxModal/NyxModal.vue'
