+++ src/composables/index.ts
@@ -1,12 +1,18 @@
+import useEditorAnnotations from './useEditorAnnotations'
+import useEditorMeta from './useEditorMeta'
 import useKeyboardShortcuts from './useKeyboardShortcuts'
 import useKeyPress from './useKeyPress'
 import useNyxProps from './useNyxProps'
 import useTeleportPosition from './useTeleportPosition'
 import { useNyxColourMode } from './useNyxColourMode'
+import { useSelectKeyboardControls } from './useSelectKeyboardControls'
 
 export {
+  useEditorAnnotations,
+  useEditorMeta,
   useKeyboardShortcuts,
   useKeyPress,
+  useSelectKeyboardControls,
   useNyxColourMode,
   useNyxProps,
   useTeleportPosition
