+++ src/utils/index.ts
@@ -1,6 +1,5 @@
 export * from './keydict'
 export * from './object'
-export { default as NyxLoader } from './NyxLoader'
 export * from './number'
 export * from './string'
 export * from './url'
